refactor test clases
