+++ 4/data/suites.csv
@@ -1,9 +1,9 @@
 "Status","Start Time","Stop Time","Duration in ms","Parent Suite","Suite","Sub Suite","Test Class","Test Method","Name","Description"
-"passed","Fri Aug 16 11:39:26 GMT 2024","Fri Aug 16 11:39:27 GMT 2024","1266","Surefire suite","Surefire test","api.ReqResInTest","api.ReqResInTest","testListResource","testListResource",""
-"passed","Fri Aug 16 11:39:28 GMT 2024","Fri Aug 16 11:39:28 GMT 2024","254","Surefire suite","Surefire test","api.ReqResInTest","api.ReqResInTest","testSingleResourceNotFound","testSingleResourceNotFound",""
+"passed","Fri Aug 16 11:39:26 GMT 2024","Fri Aug 16 11:39:27 GMT 2024","1266","Surefire suite","Surefire test","api.DeleteReqresInTest","api.DeleteReqresInTest","testListResource","testListResource",""
+"passed","Fri Aug 16 11:39:28 GMT 2024","Fri Aug 16 11:39:28 GMT 2024","254","Surefire suite","Surefire test","api.DeleteReqresInTest","api.DeleteReqresInTest","testSingleResourceNotFound","testSingleResourceNotFound",""
 "passed","Fri Aug 16 11:39:26 GMT 2024","Fri Aug 16 11:39:26 GMT 2024","517","Surefire suite","Surefire test","api.AutomationExerciseTest","api.AutomationExerciseTest","testRequestProductList","testRequestProductList",""
 "passed","Fri Aug 16 11:39:23 GMT 2024","Fri Aug 16 11:39:26 GMT 2024","2311","Surefire suite","Surefire test","api.AutomationExerciseTest","api.AutomationExerciseTest","testApiGetAllProductsList","testApiGetAllProductsList",""
-"passed","Fri Aug 16 11:39:28 GMT 2024","Fri Aug 16 11:39:28 GMT 2024","43","Surefire suite","Surefire test","api.ReqResInTest","api.ReqResInTest","testSingleUser","testSingleUser",""
-"passed","Fri Aug 16 11:39:28 GMT 2024","Fri Aug 16 11:39:28 GMT 2024","260","Surefire suite","Surefire test","api.ReqResInTest","api.ReqResInTest","testSingleResource","testSingleResource",""
-"passed","Fri Aug 16 11:39:28 GMT 2024","Fri Aug 16 11:39:28 GMT 2024","262","Surefire suite","Surefire test","api.ReqResInTest","api.ReqResInTest","testSingleUserNotFound","testSingleUserNotFound",""
-"passed","Fri Aug 16 11:39:27 GMT 2024","Fri Aug 16 11:39:28 GMT 2024","162","Surefire suite","Surefire test","api.ReqResInTest","api.ReqResInTest","testRequestListUsers","testRequestListUsers",""
+"passed","Fri Aug 16 11:39:28 GMT 2024","Fri Aug 16 11:39:28 GMT 2024","43","Surefire suite","Surefire test","api.DeleteReqresInTest","api.DeleteReqresInTest","testSingleUser","testSingleUser",""
+"passed","Fri Aug 16 11:39:28 GMT 2024","Fri Aug 16 11:39:28 GMT 2024","260","Surefire suite","Surefire test","api.DeleteReqresInTest","api.DeleteReqresInTest","testSingleResource","testSingleResource",""
+"passed","Fri Aug 16 11:39:28 GMT 2024","Fri Aug 16 11:39:28 GMT 2024","262","Surefire suite","Surefire test","api.DeleteReqresInTest","api.DeleteReqresInTest","testSingleUserNotFound","testSingleUserNotFound",""
+"passed","Fri Aug 16 11:39:27 GMT 2024","Fri Aug 16 11:39:28 GMT 2024","162","Surefire suite","Surefire test","api.DeleteReqresInTest","api.DeleteReqresInTest","testRequestListUsers","testRequestListUsers",""
