+++ 5/data/suites.csv
@@ -1,15 +1,15 @@
 "Status","Start Time","Stop Time","Duration in ms","Parent Suite","Suite","Sub Suite","Test Class","Test Method","Name","Description"
-"passed","Sun Aug 18 19:40:16 GMT 2024","Sun Aug 18 19:40:17 GMT 2024","249","Surefire suite","Surefire test","api.ReqResInTest","api.ReqResInTest","testSingleUserNotFound","testSingleUserNotFound",""
+"passed","Sun Aug 18 19:40:16 GMT 2024","Sun Aug 18 19:40:17 GMT 2024","249","Surefire suite","Surefire test","api.DeleteReqresInTest","api.DeleteReqresInTest","testSingleUserNotFound","testSingleUserNotFound",""
 "passed","Sun Aug 18 19:40:10 GMT 2024","Sun Aug 18 19:40:12 GMT 2024","2107","Surefire suite","Surefire test","api.AutomationExerciseTest","api.AutomationExerciseTest","testApiGetAllProductsList","testApiGetAllProductsList",""
-"passed","Sun Aug 18 19:40:16 GMT 2024","Sun Aug 18 19:40:16 GMT 2024","262","Surefire suite","Surefire test","api.ReqResInTest","api.ReqResInTest","testSingleResourceNotFound","testSingleResourceNotFound",""
-"passed","Sun Aug 18 19:40:16 GMT 2024","Sun Aug 18 19:40:16 GMT 2024","255","Surefire suite","Surefire test","api.ReqResInTest","api.ReqResInTest","testPutUpdateRequest","testPutUpdateRequest",""
-"passed","Sun Aug 18 19:40:15 GMT 2024","Sun Aug 18 19:40:15 GMT 2024","255","Surefire suite","Surefire test","api.ReqResInTest","api.ReqResInTest","testPatch","testPatch",""
-"passed","Sun Aug 18 19:40:13 GMT 2024","Sun Aug 18 19:40:14 GMT 2024","679","Surefire suite","Surefire test","api.ReqResInTest","api.ReqResInTest","testCreatePostPojo","testCreatePostPojo",""
-"passed","Sun Aug 18 19:40:14 GMT 2024","Sun Aug 18 19:40:14 GMT 2024","257","Surefire suite","Surefire test","api.ReqResInTest","api.ReqResInTest","testDelete","testDelete",""
-"passed","Sun Aug 18 19:40:13 GMT 2024","Sun Aug 18 19:40:13 GMT 2024","350","Surefire suite","Surefire test","api.ReqResInTest","api.ReqResInTest","testCreatePost","testCreatePost",""
-"passed","Sun Aug 18 19:40:15 GMT 2024","Sun Aug 18 19:40:15 GMT 2024","282","Surefire suite","Surefire test","api.ReqResInTest","api.ReqResInTest","testListResource","testListResource",""
-"passed","Sun Aug 18 19:40:14 GMT 2024","Sun Aug 18 19:40:15 GMT 2024","858","Surefire suite","Surefire test","api.ReqResInTest","api.ReqResInTest","testGetUsers","testGetUsers",""
-"passed","Sun Aug 18 19:40:16 GMT 2024","Sun Aug 18 19:40:16 GMT 2024","252","Surefire suite","Surefire test","api.ReqResInTest","api.ReqResInTest","testSingleResource","testSingleResource",""
+"passed","Sun Aug 18 19:40:16 GMT 2024","Sun Aug 18 19:40:16 GMT 2024","262","Surefire suite","Surefire test","api.DeleteReqresInTest","api.DeleteReqresInTest","testSingleResourceNotFound","testSingleResourceNotFound",""
+"passed","Sun Aug 18 19:40:16 GMT 2024","Sun Aug 18 19:40:16 GMT 2024","255","Surefire suite","Surefire test","api.DeleteReqresInTest","api.DeleteReqresInTest","testPutUpdateRequest","testPutUpdateRequest",""
+"passed","Sun Aug 18 19:40:15 GMT 2024","Sun Aug 18 19:40:15 GMT 2024","255","Surefire suite","Surefire test","api.DeleteReqresInTest","api.DeleteReqresInTest","testPatch","testPatch",""
+"passed","Sun Aug 18 19:40:13 GMT 2024","Sun Aug 18 19:40:14 GMT 2024","679","Surefire suite","Surefire test","api.DeleteReqresInTest","api.DeleteReqresInTest","testCreatePostPojo","testCreatePostPojo",""
+"passed","Sun Aug 18 19:40:14 GMT 2024","Sun Aug 18 19:40:14 GMT 2024","257","Surefire suite","Surefire test","api.DeleteReqresInTest","api.DeleteReqresInTest","testDelete","testDelete",""
+"passed","Sun Aug 18 19:40:13 GMT 2024","Sun Aug 18 19:40:13 GMT 2024","350","Surefire suite","Surefire test","api.DeleteReqresInTest","api.DeleteReqresInTest","testCreatePost","testCreatePost",""
+"passed","Sun Aug 18 19:40:15 GMT 2024","Sun Aug 18 19:40:15 GMT 2024","282","Surefire suite","Surefire test","api.DeleteReqresInTest","api.DeleteReqresInTest","testListResource","testListResource",""
+"passed","Sun Aug 18 19:40:14 GMT 2024","Sun Aug 18 19:40:15 GMT 2024","858","Surefire suite","Surefire test","api.DeleteReqresInTest","api.DeleteReqresInTest","testGetUsers","testGetUsers",""
+"passed","Sun Aug 18 19:40:16 GMT 2024","Sun Aug 18 19:40:16 GMT 2024","252","Surefire suite","Surefire test","api.DeleteReqresInTest","api.DeleteReqresInTest","testSingleResource","testSingleResource",""
 "passed","Sun Aug 18 19:40:12 GMT 2024","Sun Aug 18 19:40:13 GMT 2024","569","Surefire suite","Surefire test","api.AutomationExerciseTest","api.AutomationExerciseTest","testRequestProductList","testRequestProductList",""
-"passed","Sun Aug 18 19:40:15 GMT 2024","Sun Aug 18 19:40:16 GMT 2024","256","Surefire suite","Surefire test","api.ReqResInTest","api.ReqResInTest","testPutUpdate","testPutUpdate",""
-"passed","Sun Aug 18 19:40:14 GMT 2024","Sun Aug 18 19:40:14 GMT 2024","43","Surefire suite","Surefire test","api.ReqResInTest","api.ReqResInTest","testGetSingleUser","testGetSingleUser",""
+"passed","Sun Aug 18 19:40:15 GMT 2024","Sun Aug 18 19:40:16 GMT 2024","256","Surefire suite","Surefire test","api.DeleteReqresInTest","api.DeleteReqresInTest","testPutUpdate","testPutUpdate",""
+"passed","Sun Aug 18 19:40:14 GMT 2024","Sun Aug 18 19:40:14 GMT 2024","43","Surefire suite","Surefire test","api.DeleteReqresInTest","api.DeleteReqresInTest","testGetSingleUser","testGetSingleUser",""
