+++ src/test/java/api/pojo/specification/Specifications.java
@@ -1,0 +1,28 @@
+package api.pojo.specification;
+
+import io.restassured.RestAssured;
+import io.restassured.builder.RequestSpecBuilder;
+import io.restassured.builder.ResponseSpecBuilder;
+import io.restassured.http.ContentType;
+import io.restassured.specification.RequestSpecification;
+import io.restassured.specification.ResponseSpecification;
+
+public class Specifications {
+    public static RequestSpecification requestSpecification(String url) {
+        return new RequestSpecBuilder()
+                .setBaseUri(url)
+                .setContentType(ContentType.JSON)
+                .build();
+    }
+
+    public static ResponseSpecification responseSpec200() {
+        return new ResponseSpecBuilder()
+                .expectStatusCode(200)
+                .build();
+    }
+
+    public  static void installSpec(RequestSpecification requestSpecification, ResponseSpecification responseSpecification) {
+        RestAssured.requestSpecification = requestSpecification;
+        RestAssured.responseSpecification = responseSpecification;
+    }
+}
